Task #2
The logging has been performed using database secure authentication.
This has been implemented for logging using the common way in iMathCloud and for logging via Google, Github, ...
In this implementation, the two required new tables has been added to the same persistence unit than the rest of the model's tables.